journal, shared: fix warnings during compilation on 32 bits
Some filesystem magics are too big to fit in 31 bits,
and are wrapped to negative. f_type is an int on 32 bits, so
it is signed, and we get a warning on comparison.

+++ src/journal/sd-journal.c
@@ -1205,7 +1205,7 @@ static void check_network(sd_journal *j, int fd) {
                 return;
 
         j->on_network =
-                sfs.f_type == CIFS_MAGIC_NUMBER ||
+                (long)sfs.f_type == (long)CIFS_MAGIC_NUMBER ||
                 sfs.f_type == CODA_SUPER_MAGIC ||
                 sfs.f_type == NCP_SUPER_MAGIC ||
                 sfs.f_type == NFS_SUPER_MAGIC ||
